Remove the special case test for uid=0
Some package tools run configure as fakeroot.  This triggers code in
configure.ac that detects uid 0 to hardcode the installation path of the
m4 macros to the system aclocal macro dir.  This ignores any DESTDIR the
packaging tools may have set and installation fails since it attempts to
write to a system dir without proper permissions.