Prevent abort if a menu doesn't exist.
This patch prevents an abort when a nav command tries to send you to a
menu that doesn't exist. Mac the ripper's feature title extraction
removes menus from the resulting image, but does not remove navigation
instructions that attempt to jump to those menus. This patch checks that
a menu exists before acting on such  instructions. If the menu does not
exist, it puts the vm into the stopped state.