Check cell new row before using it to index into cell_playback
cellnr is used to index into cell_playback after subtracting
one from it. If cellnr is 0, then it will index -1 in cell_playback
which will seek out of boundary of cell_playback. This manifested into a
segfault for some users as reported by this launchpad bug:
https://bugs.launchpad.net/ubuntu/+source/libdvdnav/+bug/934471

By checking cellnr and skipping the indexing if cellnr is equal to zero
then we avoid the segfault. There might be a bigger issue w/ regard to
retrieving a value of zero for cell new row, but this fix works for the
reporter.


reporting the bug to launchpad and submitting a potential patch (even
though we went w/ a different fix).