Add the static function vm_close()
It is convenient to have vm_stop() have different processing when called
by other parts of the library. When vm_stop is changed in this way, it
means that the vm needs to be able to do the same processing internally.
The introduction of the static function vm_close() provides that
processing.