Check that a VOB is open in case we need to open one
If a VOB is not open we will dereference a null pointer in
DVDReadBlocks. That will produce a segfault. If we check for an open VOB
when we check to see if we need to change the currently open VOB we can
ensure that a VOB is open when DVDReadBlocks dereferences it.