Handle corrupted/badly formatted serial number data
If the serial number starts with a NULL, the strncpy may copy zero bytes,
leaving bad data in the buffer, and if the serial number is filled with data,
i-73 (where i could be 89) means one would overwrite a byte after the serial
storage.

This patch fixes both cases, and is consistent with the same approach in the
lines above copying the name.