Fix for Unknown command type handling
Some DVDs seem to use command type 7 for something.
But it will hit the 'Unknown command' check, resulting in asser(0).
I.e. DVD refuses to play.

By removing the assert() call the DVD seems to play fine.
This patch removes the assert and corrects the error message
(it is a 'command type', not a 'command')

Signed-off-by: Marcel Mol <marcel@mesa.nl>
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>